marching-squares: improve build_countours on the average case
Instead of using a `Vec` to store the segments use a `HashMap` from the
starting position of each segment with integers coordinates to a `Vec`
of segments that start in that position. Since usually there're only a
few segments that start in the same position this allows to find the
next segment of a given one in O(1) on average.